refactor: renamed word_focused to word_selection
