+++ lib/crossword/game/section_component/section_component.dart
@@ -12,7 +12,7 @@ import 'package:io_crossword/crossword/crossword.dart';
 import 'package:io_crossword/crossword/extensions/characters_rectangle.dart';
 import 'package:io_crossword/crossword/extensions/extensions.dart';
 import 'package:io_crossword/crossword/game/section_component/models/models.dart';
-import 'package:io_crossword/word_focused/word_focused.dart';
+import 'package:io_crossword/word_selection/word_selection.dart';
 
 part 'section_debug.dart';
 part 'section_keyboard_handler.dart';
