+++ test/word_focused/bloc/word_selection_state_test.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:io_crossword/word_focused/word_focused.dart';
+import 'package:io_crossword/word_selection/word_selection.dart';
 
 void main() {
   group('$WordSelectionState', () {
